extract styles  + fix collection view
