+++ src/script/components/Image/RestrictedImage.tsx
